texturess

+++ The Great Machine/Things.cpp
@@ -10,6 +10,6 @@ void Tile::Update(float fTime)
 void Tile::Draw(olc::PixelGameEngine* engine, Camera* camera)
 {
 	engine->DrawPartialSprite(
-		{ Coords.x + camera->Coords.x, Coords.y + camera->Coords.y }, 
+		{ static_cast<int>(Coords.x + camera->Coords.x), static_cast<int>(Coords.y + camera->Coords.y) }, 
 		SpriteMap, SpriteTextureMask, { 16, 16 });
 }
